Python backend/DRF Practice/requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-gis==1.0
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
